MainScreen.js: This has changes to the navigation, namely, it moves it to another class, which helps with modularity.
Results.js: This has its default class renamed for better compatibility. it was originally named flatlistsomething I dont remember properly.

ScreenNavigator.js: this .js file handles the navigation within the entire application. Update this to have navigatable screens for use in the application.

index.js: this was changed so the entry point was the new ScreenNavigator.js class instead of the MainScreen.js.